Queueable refactor
Refactors the ShoveBundlr class to be queuable.  Mostly because (duh), a single execution of ShoveBundlr quickly hits the 50K record query limit.

This refactor chains the jobs sequentially, based on the Sequence__c field on the metadata, and increments it for each subsequent job.

Finally, each object queried now saves as a discreet static resource, rather than trying to work with one monster JSON string.  This was proactively done, since a heap limit would be tripped soon after getting around the query limit.

@todo next:  write the parser (implementing SandboxPostCopy) that will insert the records stored in the destination sandbox.  I might need to work some voodoo to manage the relationships between data - but that can wait until after sleep.